requirements.txt
streamlit
pandas
scikit-learn
joblib
requests
python-dotenv
seaborn
matplotlib
plotly